backend/requirements.txt
black==24.10.0
Django==5.1.4
djangorestframework==3.15.2